backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.1.0
reportlab==4.0.7
requests==2.31.0
